fix:Correcting views widgets
